+++ src/Shashlik.DataProtection.MySql/Shashlik.DataProtection.MySql.csproj
