+++ src/Shashlik.DataProtection.Redis/Shashlik.DataProtection.Redis.csproj
