+++ src/Shashlik.sln
@@ -47,15 +47,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shashlik.Identity.MySql", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shashlik.Identity.PostgreSql", "Shashlik.Identity.PostgreSql\Shashlik.Identity.PostgreSql.csproj", "{4FCD1BF4-FD92-436F-BDB7-32853B28D054}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.DataProtector.Redis", "Shashlik.DataProtector.Redis\Shashlik.DataProtector.Redis.csproj", "{692C8050-D6F4-4566-A356-6C987F338B68}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.Response", "Shashlik.Response\Shashlik.Response.csproj", "{D23DA6B8-8104-4940-BCCF-09E45F1C9AB6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.JsonPatch", "Shashlik.JsonPatch\Shashlik.JsonPatch.csproj", "{7F170155-2445-45B2-8977-B5A671696A2C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.DataProtector.MySql", "Shashlik.DataProtector.MySql\Shashlik.DataProtector.MySql.csproj", "{EE3FBB28-BDC0-4589-91B9-DFBAABC674EA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.DataProtector.PostgreSql", "Shashlik.DataProtector.PostgreSql\Shashlik.DataProtector.PostgreSql.csproj", "{6920A36C-4D5F-4A00-A405-013F216909B2}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.Captcha.Totp", "Shashlik.Captcha.Totp\Shashlik.Captcha.Totp.csproj", "{4E706D5F-AB9A-4FC2-BDB2-7B1E99729189}"
 EndProject
@@ -72,6 +66,12 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.DataProtection.PostgreSql.Tests", "Shashlik.DataProtection.PostgreSql.Tests\Shashlik.DataProtection.PostgreSql.Tests.csproj", "{CCECB0D3-0ADC-44F4-84DD-80A1EF200449}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.JsonPatch.ModelBinder", "Shashlik.JsonPatch.ModelBinder\Shashlik.JsonPatch.ModelBinder.csproj", "{F2525FFB-D29E-4E57-B655-4F458517D7FF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.DataProtection.MySql", "Shashlik.DataProtection.MySql\Shashlik.DataProtection.MySql.csproj", "{BAA25C5A-F4BB-4C88-B071-C820EA1C696E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.DataProtection.PostgreSql", "Shashlik.DataProtection.PostgreSql\Shashlik.DataProtection.PostgreSql.csproj", "{74877BBA-41B7-45CC-ADEB-D2A2F0BC44D0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shashlik.DataProtection.Redis", "Shashlik.DataProtection.Redis\Shashlik.DataProtection.Redis.csproj", "{94ACEEEA-AFF2-44DC-84FF-397657BA9AF8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -206,12 +206,6 @@ Global
 		{4FCD1BF4-FD92-436F-BDB7-32853B28D054}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4FCD1BF4-FD92-436F-BDB7-32853B28D054}.Test|Any CPU.ActiveCfg = Debug|Any CPU
 		{4FCD1BF4-FD92-436F-BDB7-32853B28D054}.Test|Any CPU.Build.0 = Debug|Any CPU
-		{692C8050-D6F4-4566-A356-6C987F338B68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{692C8050-D6F4-4566-A356-6C987F338B68}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{692C8050-D6F4-4566-A356-6C987F338B68}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{692C8050-D6F4-4566-A356-6C987F338B68}.Release|Any CPU.Build.0 = Release|Any CPU
-		{692C8050-D6F4-4566-A356-6C987F338B68}.Test|Any CPU.ActiveCfg = Debug|Any CPU
-		{692C8050-D6F4-4566-A356-6C987F338B68}.Test|Any CPU.Build.0 = Debug|Any CPU
 		{D23DA6B8-8104-4940-BCCF-09E45F1C9AB6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D23DA6B8-8104-4940-BCCF-09E45F1C9AB6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D23DA6B8-8104-4940-BCCF-09E45F1C9AB6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -224,18 +218,6 @@ Global
 		{7F170155-2445-45B2-8977-B5A671696A2C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7F170155-2445-45B2-8977-B5A671696A2C}.Test|Any CPU.ActiveCfg = Debug|Any CPU
 		{7F170155-2445-45B2-8977-B5A671696A2C}.Test|Any CPU.Build.0 = Debug|Any CPU
-		{EE3FBB28-BDC0-4589-91B9-DFBAABC674EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE3FBB28-BDC0-4589-91B9-DFBAABC674EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EE3FBB28-BDC0-4589-91B9-DFBAABC674EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EE3FBB28-BDC0-4589-91B9-DFBAABC674EA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EE3FBB28-BDC0-4589-91B9-DFBAABC674EA}.Test|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE3FBB28-BDC0-4589-91B9-DFBAABC674EA}.Test|Any CPU.Build.0 = Debug|Any CPU
-		{6920A36C-4D5F-4A00-A405-013F216909B2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6920A36C-4D5F-4A00-A405-013F216909B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6920A36C-4D5F-4A00-A405-013F216909B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6920A36C-4D5F-4A00-A405-013F216909B2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6920A36C-4D5F-4A00-A405-013F216909B2}.Test|Any CPU.ActiveCfg = Debug|Any CPU
-		{6920A36C-4D5F-4A00-A405-013F216909B2}.Test|Any CPU.Build.0 = Debug|Any CPU
 		{4E706D5F-AB9A-4FC2-BDB2-7B1E99729189}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4E706D5F-AB9A-4FC2-BDB2-7B1E99729189}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4E706D5F-AB9A-4FC2-BDB2-7B1E99729189}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -284,6 +266,24 @@ Global
 		{F2525FFB-D29E-4E57-B655-4F458517D7FF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F2525FFB-D29E-4E57-B655-4F458517D7FF}.Test|Any CPU.ActiveCfg = Debug|Any CPU
 		{F2525FFB-D29E-4E57-B655-4F458517D7FF}.Test|Any CPU.Build.0 = Debug|Any CPU
+		{BAA25C5A-F4BB-4C88-B071-C820EA1C696E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BAA25C5A-F4BB-4C88-B071-C820EA1C696E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BAA25C5A-F4BB-4C88-B071-C820EA1C696E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BAA25C5A-F4BB-4C88-B071-C820EA1C696E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BAA25C5A-F4BB-4C88-B071-C820EA1C696E}.Test|Any CPU.ActiveCfg = Debug|Any CPU
+		{BAA25C5A-F4BB-4C88-B071-C820EA1C696E}.Test|Any CPU.Build.0 = Debug|Any CPU
+		{74877BBA-41B7-45CC-ADEB-D2A2F0BC44D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{74877BBA-41B7-45CC-ADEB-D2A2F0BC44D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{74877BBA-41B7-45CC-ADEB-D2A2F0BC44D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{74877BBA-41B7-45CC-ADEB-D2A2F0BC44D0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{74877BBA-41B7-45CC-ADEB-D2A2F0BC44D0}.Test|Any CPU.ActiveCfg = Debug|Any CPU
+		{74877BBA-41B7-45CC-ADEB-D2A2F0BC44D0}.Test|Any CPU.Build.0 = Debug|Any CPU
+		{94ACEEEA-AFF2-44DC-84FF-397657BA9AF8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{94ACEEEA-AFF2-44DC-84FF-397657BA9AF8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{94ACEEEA-AFF2-44DC-84FF-397657BA9AF8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{94ACEEEA-AFF2-44DC-84FF-397657BA9AF8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{94ACEEEA-AFF2-44DC-84FF-397657BA9AF8}.Test|Any CPU.ActiveCfg = Debug|Any CPU
+		{94ACEEEA-AFF2-44DC-84FF-397657BA9AF8}.Test|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
